requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy>=1.26.4
scikit-learn==1.8.0
joblib==1.3.2
nltk==3.8.1
requests==2.31.0
gunicorn==21.2.0
huggingface-hub==0.20.3
python-dotenv
